move BuildFrames to timeseries pkg
